Add cairo support
We need some kind of minimal rendering if we want to patch the image on
demand e.g. to add text overlays. For this purpose the cairo support is
added.

The cairo support can be enabled/disabled via ./configure switches to
not influence the existing users.

The initial cairo backend try to load the background image from an PNG
picture which can be of any size since we scale it to the size supported
by the drm device. Only the format has to match the target format. Later
on SVG support can be added as well. Once the background image is
imported it can be overlayed with an dynamic text via the
platsch_overlay_text environment variable e.g.:

  platsch_overlay_text="(300,200):text"

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>