platsch: bugfix: use double precision in cairo_scale
cairo_scale parameters are the results of an integers division.
If the imported image size is greater than surface size, the
applied scale is 0 and nothing is drawn. If the image size
is smaller than the surface size, the image drawn doesn't fit
the surface properly. Scale parameters should be double precision
to allow for the image to fit properly the surface no matter the
scale between the image and the surface.

Upstream-Status: Pending

Signed-off-by: Catalin Popescu <catalin.popescu@leica-geosystems.com>