Platsch: always fork child process
When running Platsch in the background with an ampersand,
the initramfs may disappear after the rootfs is mounted.
In such cases, Platsch may fail to read the resources.
To address this, Platsch now always forks a child process
after acquiring the resources.
The chiled process keeps the DRM devices remain open
while the parent process exits with a clear status.

Signed-off-by: LI Qingwu <Qing-wu.Li@leica-geosystems.com.cn>